+++ server/index.ts
@@ -1,3 +1,4 @@
+import "./node-als";
 import { createServer } from "node:http";
 import { parse } from "node:url";
 import next from "next";